UPSTREAM: Documentation: Remove bogus claim about del_timer_sync()
del_timer_sync() does not return the number of times it tried to delete the
timer which rearms itself. It's clearly documented:

 The function returns whether it has deactivated a pending timer or not.

This part of the documentation is from 2003 where del_timer_sync() really
returned the number of deletion attempts for unknown reasons. The code
was rewritten in 2005, but the documentation was not updated.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>
Reviewed-by: Anna-Maria Behnsen <anna-maria@linutronix.de>
Link: https://lore.kernel.org/r/20221123201624.452282769@linutronix.de
(cherry picked from commit b0b0aa5d858d4d2fe39a5e4486e0550e858108f6)

BUG=b:222168464
TEST=cq

Change-Id: Icb777cf96e82ef8a4d11da8bb64a3518ac2ee343
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4098682
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Martin Faltesek <mfaltesek@google.com>