BACKPORT: FROMGIT: drm/i915: Introduce guard pages to i915_vma
Introduce the concept of padding the i915_vma with guard pages before
and after. The major consequence is that all ordinary uses of i915_vma
must use i915_vma_offset/i915_vma_size and not i915_vma.node.start/size
directly, as the drm_mm_node will include the guard pages that surround
our object.

The biggest connundrum is how exactly to mix requesting a fixed address
with guard pages, particularly through the existing uABI. The user does
not know about guard pages, so such must be transparent to the user, and
so the execobj.offset must be that of the object itself excluding the
guard. So a PIN_OFFSET_FIXED must then be exclusive of the guard pages.
The caveat is that some placements will be impossible with guard pages,
as wrap arounds need to be avoided, and the vma itself will require a
larger node. We must not report EINVAL but ENOSPC as these are unavailable
locations within the GTT rather than conflicting user requirements.

In the next patch, we start using guard pages for scanout objects. While
these are limited to GGTT vma, on a few platforms these vma (or at least
an alias of the vma) is shared with userspace, so we may leak the
existence of such guards if we are not careful to ensure that the
execobj.offset is transparent and excludes the guards. (On such platforms
like ivb, without full-ppgtt, userspace has to use relocations so the
presence of more untouchable regions within its GTT such be of no further
issue.)

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Tejas Upadhyay <tejaskumarx.surendrakumar.upadhyay@intel.com>
Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Signed-off-by: Andi Shyti <andi.shyti@linux.intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20221201203912.346110-1-andi.shyti@linux.intel.com
(cherry picked from commit 6110225144d1136db5b026a22efbd76cee197027
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-gt-next)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:249931687
TEST=Without cleaning unused GTT pages, kernel boot time can improve at ~600ms
     on ADL-P celeron sku.

Change-Id: Ic725f802cf6d162d546e73e3b4e93a06d9f4a5a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4075452
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Andi <andi.shyti@gmail.com>
Tested-by: Andi <andi.shyti@gmail.com>
Tested-by: Shawn C Lee <shawn.c.lee@intel.com>