FROMLIST: drm/amd/display: Fix dsc mismatch of acquire and validation of dsc engine
[Why]
We skip dsc_validation on pipes that are underlays, but in the
acquire_dsc code we don't have this check.

In certain conditions (when underlay pipe index is lower) we will assign
the dsc resource to the underlay pipe and skip the base pipe.

Now during dsc_validation we will skip the underlay pipe (this has the
dsc resource) but try to validate the base pipe(this doesn't have a dsc
resource) due to this mismatch we hit a NULLPTR

[How]
In the acquire_dsc add a check for underlay pipe so we
don't acquire a dsc resource for this pipe. This will match the
acquire/validation conditions.

Reviewed-by: Wenjing Liu <Wenjing.Liu@amd.com>
Reviewed-by: Hersen Wu <Hersenxs.Wu@amd.com>
Acked-by: Praful Swarnakar <Praful.Swarnakar@amd.com>
Signed-off-by: Bhawanpreet Lakha <Bhawanpreet.Lakha@amd.com>

(am from https://www.spinics.net/lists/amd-gfx/msg86833.html)
(also found at https://lore.kernel.org/all/20221221184359.546013-1-Bhawanpreet.Lakha@amd.com/)

BUG=b:240216859
TEST=Build and boot kernel, perform Video playback with lid open/close

Change-Id: Ie7f35b3ccc5e4b8ac01e47aba68d42ec11b51120
Signed-off-by: Praful Swarnakar <praful.swarnakar@amd.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4127623
Reviewed-by: Raul Rangel <rrangel@chromium.org>