xhci-exynos: Maintain a list of udev requiring autosuspend
xhci-exynos requires the udev to set the remote_wakeup bit in the
descriptor to enable autosuspend feature. However, some accessories do
not report remote_wakeup capability to the Host and still work well if
the Host suspends the bus. Maintain a list of the VID:PID pair of such
accessories so that xhci-exynos is able to autosuspend when these
accessories are enumerated.

Bug: 254151227
Change-Id: Ief3e67a19471d8cc730e2792ea44797c76957c4a
Signed-off-by: Kyle Tso <kyletso@google.com>
(cherry picked from commit 3becf0c3ffaa1e96edac962f280237a8467d3f8c)