CHROMIUM: config: reven: configure psmouse driver as a kernel module
The psmouse ALPS driver extension intermittently fails
initialization because it times out on receiving responses
from the ALPS touchpad on the Dell Vostro 3550. This results in
the kernel falling back to the generic PS/2 mouse driver and the
touchpad having only basic pointer functionality. By configuring
the psmouse driver as a kernel module and not as built-in, the
driver timing is more stable, the ALPS extension initializes
correctly and the touchpad has full functionality.

BUG=b:258477982
TEST=./chromeos/scripts/kernelconfig olddefconfig
TEST=`USE=dyndebug emerge-reven chromeos-kernel-5_10` &&
`update_kernel.sh --remote=$IP --ssh_port 22`
TEST=verify CONFIG_MOUSE_PS2=m and vertical scrolling works

Change-Id: Iaaa7202d748d7b96a630afb74085a0e90e030f60
Signed-off-by: Jonathan Denose <jdenose@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4245697
Reviewed-by: Jeffery Miller <jefferymiller@google.com>