CHROMIUM: usb: core: Add error reporting to the hub driver
Add uevent based error reporting to drivers/usb/core/hub.c. Have
functions in this driver which are responsible for device enumeration
and downstream port management send a uevent with an error code when
the function returns an error.

BUG=b:269045190
TEST=emerge-brya chromeos-kernel-5_10. Also, manually tested uevents
are received in userspace on hub_report_error calls using a brya
with a custom v5.10 kernel image including this change.
UPSTREAM-TASK=b:269131835

Signed-off-by: Jameson Thies <jthies@google.com>
Change-Id: Ie82b8057d48b7a281d7b95ee9ed0f3cbc35dc654
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4246280
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>