CHROMIUM: usb: host: Add error reporting to the xHCI driver
Add uevent based error reporting to drivers/usb/host/xhci.c. Have
functions in this driver which are responsible for setting up devices
as well as endpoint, bandwidth and stream management send a uevent
with an error code when the function returns an error.

BUG=b:269045190
TEST=emerge-brya chromeos-kernel-5_10. Also, manually tested uevents
are received in userspace on xhci_report_error calls using a brya
with a custom v5.10 kernel image including this change.
UPSTREAM-TASK=b:269131835

Signed-off-by: Jameson Thies <jthies@google.com>
Change-Id: Ia88896955bd0daffd01bba75650f74287f944e37
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4246284
Reviewed-by: Benson Leung <bleung@google.com>