CHROMIUM: iwl7000: mac80211: make casts more typesafe and more generic
We have a some helpers that just pass around a void pointer paired
with macros that hook this up to work in different contexts with
specific types. However, instead of doing type checking, we just
casted the function pointers to take a (void *) as the argument. It
appears that doing so causes compilation failures with
-Wcast-function-type-strict in some cases.

So, instead first cast it to the correct exact type that is expected
(so the compiler does a typesafety check) and then cast the function
to (void *) to avoid any possibility of the compiler warning about
the cast.

Change-Id: Idfddb2076b6b7c9f3dffeb6e643a4a3492950feb
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 2139c7bba4a710d8bf7f2b8bb00c0224f8c13db2