CHROMIUM: iwl7000: mvm: increase debug level when we fail to associate
Chrome users report a high number of "Not associated and the time event
is over already..." errors.
By looking at the logs it seems that we can't get the authentication to
go out, or at least, we don't hear the authentication response.
Add more logs to debug this area specifically.
This is a temporary patch to debug this specific issue that should be
shipped to the users and the code is written in such a way that it'll be
easy to ship it to Chrome users only: under an ifdef.
This is also the reason the copyright is not changed: it makes little
sense to change the copyright for a patch that will be reverted.

BUG=b:270746800
TEST=basic functionality and log collection

Change-Id: I09d5a0983adf88c426b234eecf05a3b2bd768a1d
iwl7000-tree: 88b2d102311c53e90ec07566b6bf7e1597c1e55d
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4338181
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>