CHROMIUM: security: DM LSM to check uuid instead of name
Device Mapper names are mutable, this means that mknod could be
accomplished by renaming a dm device. UUIDs however cannot be changed.
This CL will check the DM device for a dm_locked- prefix, and if
it's set mknod will be rejected.

When a dm device is created if we set the uuid to dm_locked-NAME then
any future mknods will be rejected, regardless of the device name.

	~ # dmsetup info hiberimage2
	Name:              hiberimage2
	State:             ACTIVE
	Read Ahead:        512
	Tables present:    LIVE
	Open count:        0
	Event number:      2
	Major, minor:      254, 10
	Number of targets: 1
	UUID: dm_locked-hiberimage-foo-bar

	~ # dmsetup mknodes hiberimage2
	/dev/mapper/hiberimage2: mknod for hiberimage2 failed: Operation not permitted

UPSTREAM-TASK=b:273250069
TEST=Brya-hibernate manually
BUG=b:266841303

Change-Id: I9caf722ee4cc94f15cb388edc6b1c2c095142555
Signed-off-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4346695
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>