CHROMIUM: add cros specific trace-points
!!!DO NOT UPSTREAM!!!
!!!DO NOT PORT TO OTHER KERNELS!!!

Add a tracepoint that will monitor socket listens.

The rationale is that fentry/fexit bpf hooks for ARM64 are not available
v5.15 kernel. Instead of cherry-picking many, many patches to support
ARM64 fentry/fexit hooks, we add this tracepoint.

This patch is identical to the patch reviewed in 4429631 except that it
is targeted at the 5.10 kernel.

BUG=b:251499575
TEST=compiled and tested on arm64-generic, attached bpf to the
tracepoint and verified correct information was being received in
userspace.
UPSTREAM-TASK=b:278303728

Change-Id: Ib1a670d2b543969409a6def4595923234cd7da47
Signed-off-by: Jason Ling <jasonling@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4518626
Reviewed-by: Aashay Shringarpure <aashay@google.com>