devfreq: Use lower bound for memlat/dsulat
The previous logic looks up the target frequency for the first node
*greater or equal* to the current frequency.
This results in an awkward lookup logic (both in code, and when creating
or reading frequency lookup tables).
We should instead use the target frequency for the highest node
with a frequency *less or equal* to the current frequency.
I.e. the table entries should each represent new explicit thresholds
for frequencies exceeding the specified values.

Bug: 287102956
Change-Id: Id4c0436e46e62e401a96907a3f9b4f78511943b5
Signed-off-by: Vincent Palomares <paillon@google.com>