Add a debugfs node to refresh slc_pmon
slc_pmon module only loads the event list once when it initializes.
On P21/P22, to refresh the event list, we load/unload the module
after the privileged debug mode is enabled. However, on P23, we cannot
use the temporary solution because slc_pmon module is the part of the
initramfs image. So, we add a debugfs node to refresh the event list for
slc_pmon module.

node path: /sys/kernel/debug/slc_pmon/refresh

Write anything to the node, the event list for the slc_pmon reload.

Bug: 289429421
Change-Id: Ic45d9b7a8f9fca4168d113af6df4f57fdb7ae8b5
Signed-off-by: Taeju Park <taeju@google.com>