tcpci_max77759: Debounce adapter disconnects
In certain hard to reproduce scenarios, adapter seems to drop VBUS
and/or disconnect for brief period of time after which the connection
recovers. This could possibly be due to CHGIN overshoot. Debounce
disconnects for a second when last negotiated current limit is
greater than 1.5A. Drop the current limit to 500mA while the
connection recovers. This mitigates the back to back notifications
to the userspace when the adapter disconnects and recovers within
a brief amount of time. On the flip side, the phone will take ~1 more
second to update the UX.

Bug: 294875522
Test: Manually tested for usb online updates upon connnect/disconnect
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>
Change-Id: Iab22850e91b8b062b63aa4115b1d22f54ebe786b